requirements.txt
python-dotenv~=0.19.2
requests~=2.27.0
pymongo~=3.12.3
setuptools~=57.0.0
mongoengine~=0.23.1
python-dateutil~=2.8.2
torch~=1.10.1
matplotlib~=3.5.1
plotly~=5.5.0